% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_streetview.R
\name{browse_streetview}
\alias{browse_streetview}
\title{Browse Street View}
\usage{
browse_streetview(parcels, title = "Street View", limit = 50)
}
\arguments{
\item{parcels}{A data frame containing parcel information, including an \code{Address} column.}

\item{title}{Optional. A single string for the app's title. Defaults to \code{"Street View"}.}

\item{limit}{Optional. A numeric value to limit the number of parcels viewed. Defaults to 50.}
}
\value{
A Shiny app is launched to browse street views, using the given parcels.
}
\description{
A short description...
}
