% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer}
\alias{buffer}
\title{Create a buffer around geometry}
\usage{
buffer(geometry, radius = 1, center = alico, clean = F)
}
\arguments{
\item{geometry}{A geometric object.}

\item{radius}{Optional. A numeric radius, defaulting to 1.}

\item{center}{Optional. A geometric center, defaulting to \code{alico}.}

\item{clean}{Optional. A logical value indicating whether to cleanly cut the buffer, defaulting to \code{FALSE}.}
}
\value{
A buffered geometric object, either intersected with \code{geometry} for a clean result or retaining full parcels or geometry otherwise.
}
\description{
A short description...
}
