% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_standard.R
\name{filter_standard}
\alias{filter_standard}
\title{Standard filter on geometry}
\usage{
filter_standard(geometry)
}
\arguments{
\item{geometry}{A spatial object.}
}
\value{
A filtered spatial object with additional area calculations. The object is filtered
to include only those with an area greater than 1000 square feet, exclude certain
zoning types, and exclude certain land use plans.
}
\description{
A short description...
}
