% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_places.R
\name{get_places}
\alias{get_places}
\title{Get places around a parcel}
\usage{
get_places(parcel, radius = 100, first = TRUE, wholetable = FALSE)
}
\arguments{
\item{parcel}{A spatial object representing the parcel.}

\item{radius}{A numeric value specifying the search radius in meters. Optional.}

\item{first}{Optional. A logical value.}

\item{wholetable}{Optional. A logical value.}
}
\value{
A data frame of places within the parcel. If \code{first} is TRUE, only the first result is returned. If no places are found and \code{wholetable} is FALSE, a tibble with "none" values is returned.
}
\description{
A short description...
}
