% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_long.R
\name{lat_long}
\alias{lat_long}
\title{Extract latitude and longitude}
\usage{
lat_long(parcel)
}
\arguments{
\item{parcel}{A spatial object representing a parcel.}
}
\value{
A named numeric vector with \code{latitude} and \code{longitude}.
}
\description{
A short description...
}
