% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_infrastructure.R
\name{map_infrastructure_cost}
\alias{map_infrastructure_cost}
\title{Map Infrastructure Cost}
\usage{
map_infrastructure_cost(
  data,
  map_center = NULL,
  zoom = 12,
  pitch = 80,
  bearing = 270,
  legend_title = NULL
)
}
\arguments{
\item{data}{A dataset, likely a spatial object.}

\item{map_center}{Optional. A numeric vector representing the map's center coordinates.}

\item{zoom}{Optional. A numeric value for the zoom level.}

\item{pitch}{Optional. A numeric value for the pitch.}

\item{bearing}{Optional. A numeric value for the bearing.}

\item{legend_title}{Optional. A title for the legend.}
}
\value{
A map visualization. The function will instead raise an error if required columns aren't present in the data.
}
\description{
A short description...
}
