% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_value.R
\name{map_value}
\alias{map_value}
\title{Map value}
\usage{
map_value(
  data,
  value_column = "val_per_acre",
  tooltip_column = NULL,
  map_center = NULL,
  zoom = 12,
  pitch = 80,
  bearing = 270,
  legend_title = NULL
)
}
\arguments{
\item{data}{A spatial data object.}

\item{value_column}{A single string. Optional. Default is 'val_per_acre'.}

\item{tooltip_column}{A single string. Optional.}

\item{map_center}{A numeric vector of length 2. Optional.}

\item{zoom}{A single numeric value. Optional.}

\item{pitch}{A single numeric value. Optional.}

\item{bearing}{A single numeric value. Optional.}

\item{legend_title}{A single string. Optional.}
}
\value{
The function generates a map visualizing the data.
}
\description{
A short description...
}
