% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_common_words.R
\name{remove_common_words}
\alias{remove_common_words}
\title{Remove Common Words from a Dataframe Column}
\usage{
remove_common_words(df, column, top_n = 100)
}
\arguments{
\item{df}{A dataframe.}

\item{column}{A column in \code{df} to process, captured as a symbol.}

\item{top_n}{Optional. A numeric value indicating the number of top common words to remove, defaulting to 100.}
}
\value{
The dataframe with an additional column containing the processed text, with the most common words removed.
}
\description{
A short description...
}
