% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_cost.R
\name{road_cost}
\alias{road_cost}
\title{Calculate road maintenance cost}
\usage{
road_cost(road, annualized = T, measure_length = T)
}
\arguments{
\item{road}{An \code{sf} object representing the road.}

\item{annualized}{A logical value. If \code{TRUE}, returns the annualized maintenance cost; otherwise, returns the short-term cost. Optional.}

\item{measure_length}{A logical value. If \code{FALSE}, uses the \code{SECTLGTH} for calculation instead of the road length. Optional.}
}
\value{
The calculated cost, either annualized or short-term, as a numeric value.
}
\description{
A short description...
}
