% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_return.R
\name{road_return}
\alias{road_return}
\title{Calculate road return values}
\usage{
road_return(
  roads,
  per_sqft = T,
  annualized = T,
  just_value = F,
  just_revenue = F
)
}
\arguments{
\item{roads}{A spatial object containing roads.}

\item{per_sqft}{Optional. A logical value indicating whether to return values per square foot. Defaults to \code{TRUE}.}

\item{annualized}{Optional. A logical value indicating whether the costs should be annualized. Defaults to \code{TRUE}.}

\item{just_value}{Optional. A logical value. If \code{TRUE}, the function returns only the surrounding values. Defaults to \code{FALSE}.}

\item{just_revenue}{Optional. A logical value. If \code{TRUE}, the function returns only the surrounding revenues. Defaults to \code{FALSE}.}
}
\value{
A numeric vector representing the differences between surrounding revenues and road costs for each road. If \code{per_sqft} is \code{TRUE}, values are per square foot. The function returns only surrounding values if \code{just_value} is \code{TRUE}, or only surrounding revenues if \code{just_revenue} is \code{TRUE}.
}
\description{
A short description...
}
