% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agendas.R
\name{rss_update}
\alias{rss_update}
\title{Update RSS Feed}
\usage{
rss_update(rss_url)
}
\arguments{
\item{rss_url}{A single string representing the URL of the RSS feed.}
}
\value{
A tibble containing the RSS feed data with documents extracted from descriptions.
}
\description{
Waco uses IQM2 for hosting meeting agendas, etc.
This is being retired in 2027, so they will move on to another service. Hopefully Legistar,
because there's already packages for scraping that
}
