% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrounding_roads.R
\name{surrounding_roads}
\alias{surrounding_roads}
\title{Surrounding Roads}
\usage{
surrounding_roads(
  parcel,
  return = c("area", "annual cost", "repair cost", "everything", "table"),
  buffer_distance = 75
)
}
\arguments{
\item{parcel}{A parcel represented as an sf object.}

\item{return}{One of \code{'area'}, \code{'annual cost'}, \code{'repair cost'}, \code{'everything'}, or \code{'table'}.}

\item{buffer_distance}{A numeric value representing buffer distance. Optional, defaults to 75.}
}
\value{
Depending on the \code{return} parameter:
\itemize{
\item \code{'area'}: Total area of intersecting roads.
\item \code{'annual cost'}: Total annual cost of intersecting roads.
\item \code{'repair cost'}: Total repair cost of intersecting roads.
\item \code{'everything'}: A summary table.
\item \code{'table'}: An sf object of intersecting roads.
}
}
\description{
A short description...
}
